# backend/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "BMSTU Web Course 2020 project",
    "title": "Forums API",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "paths": {
...
    "/api/v1/thread/:slug/details": {
      "get": {
        "description": "Получение информации о ветке обсуждения по его имени.",
        "tags": [
          "threads"
        ],
        "summary": "Получение информации о ветке обсуждения",
        "operationId": "threadGetOne",
        "parameters": [
          {
            "type": "string",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветке обсуждения.",
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      },
      "post": {
        "description": "Обновление ветки обсуждения на форуме.",
        "tags": [
          "threads"
        ],
        "summary": "Обновление ветки",
        "operationId": "threadUpdate",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Данные ветки обсуждения.",
            "name": "thread",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ThreadUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветке обсуждения.",
            "schema": {
              "$ref": "#/definitions/Thread"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
    "/api/v1/user/:nickname/create": {
      "post": {
        "description": "Создание нового пользователя в базе данных.",
        "tags": [
          "user"
        ],
        "summary": "Создание нового пользователя",
        "operationId": "userCreate",
        "parameters": [
          {
            "type": "string",
            "description": "Идентификатор пользователя.",
            "name": "nickname",
            "in": "path",
            "required": true
          },
          {
            "description": "Данные пользовательского профиля.",
            "name": "profile",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Пользователь успешно создан. Возвращает данные созданного пользователя.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "409": {
            "description": "Пользователь уже присутсвует в базе данных. Возвращает данные ранее созданных пользователей с тем же nickname-ом иои email-ом.",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/api/v1/user/:nickname/profile": {
      "get": {
        "description": "Получение информации о пользователе форума по его имени.",
        "tags": [
          "user"
        ],
        "summary": "Получение информации о пользователе",
        "operationId": "userGetOne",
        "parameters": [
          {
            "type": "string",
            "description": "Идентификатор пользователя.",
            "name": "nickname",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Пользователь отсутсвует в системе.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      },
      "post": {
        "description": "Изменение информации в профиле пользователя.",
        "tags": [
          "user"
        ],
        "summary": "Изменение данных о пользователе",
        "operationId": "userUpdate",
        "parameters": [
          {
            "type": "string",
            "description": "Идентификатор пользователя.",
            "name": "nickname",
            "in": "path",
            "required": true
          },
          {
            "description": "Изменения профиля пользователя.",
            "name": "profile",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Актуальная информация о пользователе после изменения профиля.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Пользователь отсутсвует в системе.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          },
          "409": {
            "description": "Новые данные профиля пользователя конфликтуют с имеющимися пользователями.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/api/v1/user/auth": {
      "post": {
        "description": "Авторизая пользователя, возвращает cookie c токеном.",
        "tags": [
          "user"
        ],
        "summary": "Авторизация пользователя",
        "operationId": "userAuth",
        "parameters": [
          {
            "description": "Данные пользователя.",
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLogIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь успешно авторизован. Устанавливается HttpOnly cookie Auth с токеном"
          },
          "403": {
            "description": "Неверные данные пользователя"
          }
        }
      }
    },
    "/api/v1/user/check": {
      "get": {
        "description": "Проверка авторизации пользователя, возвращает информацию о пользователе, если он авторизован",
        "tags": [
          "user"
        ],
        "summary": "Проверка авторизации пользователя",
        "operationId": "userCheck",
        "responses": {
          "200": {
            "description": "Пользователь авторизован",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "403": {
            "description": "Пользователь не был авторизован"
          }
        }
      }
    },
    "/api/v1/user/signout": {
      "get": {
        "description": "Разлогирование пользователя",
        "tags": [
          "user"
        ],
        "summary": "Разлогирование пользователя",
        "operationId": "signOut",
        "responses": {
          "200": {
            "description": "Cookie Auth проставляется просроченной"
          },
          "400": {
            "description": "Пользователь не был авторизован"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Thread": {
      "description": "Ветка обсуждений на форуме",
      "type": "object",
      "properties": {
        "author": {
          "description": "Пользователь, создавший ветку\n\nformat: identity",
          "type": "string",
          "x-go-name": "Author",
          "example": "j.sparrow"
        },
        "created": {
          "description": "Дата создания ветки",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created",
          "example": "2017-01-01T00:00:00.000Z"
        },
        "forum": {
          "description": "Форум ветки\n\nformat: identity",
          "type": "string",
          "x-go-name": "ForumName",
          "readOnly": true,
          "example": "pirate-stories"
        },
        "id": {
          "description": "Идентификатор ветки обсуждения",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id",
          "readOnly": true,
          "example": 42
        },
        "message": {
          "description": "Описание ветки обсуждения",
          "type": "string",
          "x-go-name": "Message",
          "example": "An urgent need to reveal the hiding place of Davy Jones. Who is willing to help in this matter?"
        },
        "posts": {
          "description": "Количество сообщений в ветке",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PostsCount",
          "readOnly": true,
          "example": 100
        },
        "slug": {
          "description": "Человекопонятный URL\n\nformat: identity",
          "type": "string",
          "x-go-name": "Slug",
          "readOnly": true,
          "example": "jones-cache"
        },
        "title": {
          "description": "Заголовок ветки",
          "type": "string",
          "x-go-name": "Title",
          "example": "Davy Jones cache"
        },
        "votes": {
          "description": "Кол-во голосов",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Votes",
          "readOnly": true
        }
      },
      "x-go-package": "dbProject/forum/delivery/http"
    },
    "ThreadUpdate": {
      "type": "object",
      "title": "Сообщение для обновления ветки обсуждения на форуме. Пустые параметры остаются без изменений.",
      "properties": {
        "message": {
          "description": "Описание ветки обсуждения",
          "type": "string",
          "x-go-name": "Message",
          "example": "An urgent need to reveal the hiding place of Davy Jones. Who is willing to help in this matter?"
        },
        "title": {
          "description": "Заголовок ветки обсуждения",
          "type": "string",
          "x-go-name": "Title",
          "example": "Davy Jones cache"
        }
      },
      "x-go-package": "dbProject/forum/delivery/http"
    },
...
    "User": {
      "description": "Информация о пользователе",
      "type": "object",
      "properties": {
        "about": {
          "description": "Описание пользователя",
          "type": "string",
          "x-go-name": "About",
          "example": "This is the day you will always remember as the day that you almost caught Captain Jack Sparrow!"
        },
        "email": {
          "description": "Почтовый адрес пользователя\n\nformat: email",
          "type": "string",
          "x-go-name": "Email",
          "example": "captaina@blackpearl.sea"
        },
        "fullname": {
          "description": "Полное имя пользователя",
          "type": "string",
          "x-go-name": "Fullname",
          "example": "Captain Jack Sparrow"
        },
        "nickname": {
          "description": "Имя пользователя (уникальное поле)\n\nformat: identity",
          "type": "string",
          "x-go-name": "Nickname",
          "readOnly": true,
          "example": "j.sparrow"
        },
        "password": {
          "description": "Пароль пользователя",
          "type": "string",
          "x-go-name": "Password",
          "example": "123456"
        }
      },
      "x-go-name": "userInput",
      "x-go-package": "dbProject/user/delivery/http"
    },
    "UserLogIn": {
      "description": "Данные пользователя для авторизации",
      "type": "object",
      "properties": {
        "nickname": {
          "description": "Имя пользователя",
          "type": "string",
          "x-go-name": "Nickname"
        },
        "password": {
          "description": "Пароль пользователя",
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-name": "signInInput",
      "x-go-package": "dbProject/user/delivery/http"
    },
    "UserOutput": {
      "type": "object",
      "properties": {
        "about": {
          "type": "string",
          "x-go-name": "About"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "fullname": {
          "type": "string",
          "x-go-name": "Fullname"
        },
        "nickname": {
          "type": "string",
          "x-go-name": "Nickname"
        }
      },
      "x-go-package": "dbProject/user/delivery/http"
    },
    "Users": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserOutput"
          },
          "x-go-name": "Users"
        }
      },
      "x-go-package": "dbProject/user/delivery/http"
    },
...
  },
  "responses": {
    "Error": {
      "description": "Error response payload",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "Текстовое описание ошибки.",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}